Set uncompressed size in local file header to 0
The ZIP spec [1] describes in section 4.4.9 that the uncompressed size
field is set to zero in the local file header if bit 3 of the general
purpose flags is set, which zipline does.

Some applications actually care about this, e.g. Powerpoint regards a
file as corrupt when the uncompressed size and bit 3 are set.

[1] https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT